projectcopy/atc_ai_project/requirements_opensky.txt
# OpenSky Network API Integration Requirements

# Core dependencies
requests>=2.31.0
pydantic>=2.0.0
tenacity>=8.2.0
cachetools>=5.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Async support (optional but recommended)
aiohttp>=3.8.0

# Visualization (for examples)
matplotlib>=3.7.0
plotly>=5.15.0

# Excel export support (optional)
openpyxl>=3.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Code quality
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0

# Environment management
python-dotenv>=1.0.0